requirements.txt
python-telegram-bot[job-queue]==20.7
httpx==0.27.0
starlette==0.37.2
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1